+++ matrixFuncs/matrixTranspose.cpp
@@ -2,7 +2,7 @@
 #include "../linker.h"
 using namespace std;
 
-void transposeExt(float inputArr[5][5]){
+void transposeExt(double inputArr[5][5]){
   for(int i=0; i<5; i++){
     cout << "[";
     for(int j=0; j<5; j++){
